Complete code for scraping name and population of the countries
I have used and implemented selenium for webscraping data for the names and populations all over the world. I have implemented the complete correct code for moving ahead to scrap dynamic data. I was facing problem in using X_PATH but finally got to know the correct syntax for the same.